bit field in ChatMember

+++ include/telebot.h
@@ -55,7 +55,7 @@
 
 
 /* type bool */
-typedef short bool;
+typedef int bool;
 #define true  1
 #define false 0
